Move get_user_environment_variables to environ.py
